refactor: reorganizes badge variants and helper functions so reused ones are kept in variants file, but storage specific ones are moved to a storage specific utils file.
